Move TextHighlight component to its own file
